Support for creating overview by work type
Support for creating csv files with details on time spent per type of work.
This first implementation retrieves the worklogs and issues again, similar as with the
overview by epic. It would be an improvement to get the worklogs and issues only once and do both
exports with the same objects to reduce JIRA API usage.